requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
enum34==1.1.10
exceptiongroup==1.2.0
future==0.18.3
gkeepapi==0.14.2
gpsoauth==1.0.2
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pycryptodomex==3.19.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.2
trio-websocket==0.11.1
undetected-chromedriver==3.5.4
urllib3==1.26.16
websockets==12.0
wsproto==1.2.0
